pyproject.toml
[project]
name = "lastautai"
version = "0.1.0"
description = "Natural Language to Workflow Automation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.9.0",
    "pyyaml>=6.0",
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "jsonschema>=4.20.0",
]

[project.scripts]
lastautai = "src.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "golden: Golden test suite",
]
asyncio_mode = "auto"
